+++ packages/libs/contracts/src/integrations/openbanking/models.ts
@@ -97,3 +97,4 @@ export const AccountBalanceRecord = new SchemaModel({
 
 
 
+