+++ packages/libs/presentation-runtime-core/tsdown.config.js
@@ -3,11 +3,6 @@ import { moduleLibrary } from '@lssm/tool.tsdown';
 
 export default defineConfig((options) => ({
   ...moduleLibrary,
-  exports: {
-    all: true,
-    devExports: true,
-  },
-  skipNodeModulesBundle: true,
 }));