+++ packages/libs/accessibility/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @lssm/lib.accessibility
 
+## 1.7.4
+
+### Patch Changes
+
+- fix typing
+- Updated dependencies
+  - @lssm/lib.design-system@1.7.4
+  - @lssm/lib.ui-kit-web@1.7.4
+  - @lssm/lib.ui-kit@1.7.4
+
+## 1.7.3
+
+### Patch Changes
+
+- add right-sidebar
+- Updated dependencies
+  - @lssm/lib.design-system@1.7.3
+  - @lssm/lib.ui-kit@1.7.3
+  - @lssm/lib.ui-kit-web@1.7.3
+
 ## 1.7.2
 
 ### Patch Changes