+++ packages/libs/bus/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @lssm/lib.bus
 
+## 1.7.4
+
+### Patch Changes
+
+- fix typing
+- Updated dependencies
+  - @lssm/lib.contracts@1.7.4
+  - @lssm/lib.schema@1.7.4
+
+## 1.7.3
+
+### Patch Changes
+
+- add right-sidebar
+- Updated dependencies
+  - @lssm/lib.contracts@1.7.3
+  - @lssm/lib.schema@1.7.3
+
 ## 1.7.2
 
 ### Patch Changes