+++ packages/libs/contracts/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @lssm/lib.contracts
 
+## 1.7.4
+
+### Patch Changes
+
+- fix typing
+- Updated dependencies
+  - @lssm/lib.schema@1.7.4
+
+## 1.7.3
+
+### Patch Changes
+
+- add right-sidebar
+- Updated dependencies
+  - @lssm/lib.schema@1.7.3
+
 ## 1.7.2
 
 ### Patch Changes