+++ packages/libs/presentation-runtime-react-native/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @lssm/lib.presentation-runtime-react-native
 
+## 8.0.4
+
+### Patch Changes
+
+- fix typing
+- Updated dependencies
+  - @lssm/lib.presentation-runtime-core@1.7.4
+  - @lssm/lib.ui-kit@1.7.4
+
+## 8.0.3
+
+### Patch Changes
+
+- add right-sidebar
+- Updated dependencies
+  - @lssm/lib.presentation-runtime-core@1.7.3
+  - @lssm/lib.ui-kit@1.7.3
+
 ## 8.0.2
 
 ### Patch Changes