+++ packages/tools/tsdown/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @lssm/tool.tsdown
 
+## 0.8.4
+
+### Patch Changes
+
+- fix typing
+
+## 0.8.3
+
+### Patch Changes
+
+- add right-sidebar
+
 ## 0.8.2
 
 ### Patch Changes