+++ packages/tools/typescript/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @lssm/tool.typescript
 
+## 0.7.4
+
+### Patch Changes
+
+- fix typing
+
+## 0.7.3
+
+### Patch Changes
+
+- add right-sidebar
+
 ## 0.7.2
 
 ### Patch Changes